requirements.txt
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
requests==2.31.0
lxml==5.2.2
